# Conflicts:
#	app/src/main/java/com/techswivel/qthemusic/dataManager/RemoteDataManager.kt
#	app/src/main/java/com/techswivel/qthemusic/dataManager/RemoteDataManagerImp.kt
#	app/src/main/java/com/techswivel/qthemusic/ui/activities/authActivity/AuthActivity.kt
#	app/src/main/java/com/techswivel/qthemusic/ui/activities/authActivity/AuthActivityImp.kt
#	app/src/main/java/com/techswivel/qthemusic/ui/activities/authActivity/AuthActivityViewModel.kt
#	app/src/main/java/com/techswivel/qthemusic/ui/activities/mainActivity/MainActivity.kt
#	app/src/main/java/com/techswivel/qthemusic/ui/activities/profileSettingScreen/ProfileSettingActivityImpl.kt
#	app/src/main/java/com/techswivel/qthemusic/ui/base/BaseViewModel.kt
#	app/src/main/java/com/techswivel/qthemusic/ui/fragments/otpVerificationFragment/OtpVerification.kt
#	app/src/main/java/com/techswivel/qthemusic/ui/fragments/otpVerificationFragment/OtpVerificationViewModel.kt
#	app/src/main/java/com/techswivel/qthemusic/ui/fragments/profileSettingFragment/ProfileSettingFragment.kt
#	app/src/main/java/com/techswivel/qthemusic/utils/CommonKeys.kt
#	app/src/main/res/color/radio_button_color.xml
#	app/src/main/res/drawable/border_white.xml
#	app/src/main/res/drawable/ic_icon_ionic_ios_arrow_down.xml
#	app/src/main/res/values/colors.xml
#	app/src/main/res/values/strings.xml
#	build.gradle

+++ app/src/main/java/com/techswivel/qthemusic/models/AuthModel.kt
@@ -13,7 +13,7 @@ data class AuthModel(
     @SerializedName("jwt")
     var jwt: String?,
     @SerializedName("dOB")
-    var dOB: Int?,
+    var dOB: Long?,
     @SerializedName("phoneNumber")
     var phoneNumber: String?,
     @SerializedName("gender")
